requirements.txt
# Для работы с MySQL
pymysql

# Для работы с MongoDB
pymongo

#Для загрузки из файла .env в окружение os.getenv('MONGO_HOST'
python-dotenv
